@@ -3,29 +3,28 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start:digitaltesten": "ng serve --configuration=digitaltesten",
-    "start:muntligtesten": "ng serve --configuration=muntligtesten",
-    "build:prod:digitaltesten": "ng build --configuration=production-digitaltesten",
-    "build:prod:muntligtesten": "ng build --configuration=production-muntligtesten",
-    "build:prod": "npm run build:prod:digitaltesten && npm run build:prod:muntligtesten",
+    "start": "ng serve",
+    "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "format:check": "prettier --config ./.prettierrc --list-different 'src/{app,environments,assets,i18n,styles}/**/*{.ts,.js,.json,.css,.scss}'",
+    "format": "prettier --config ./.prettierrc --write 'src/{app,environments,assets,i18n,styles}/**/*{.ts,.js,.json,.css,.scss}'"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.1.2",
+    "@angular/animations": "^10.2.0",
     "@angular/cdk": "^10.2.1",
-    "@angular/common": "^10.1.2",
-    "@angular/compiler": "^10.1.2",
-    "@angular/core": "^10.1.2",
-    "@angular/forms": "^10.1.2",
-    "@angular/platform-browser": "^10.1.2",
-    "@angular/platform-browser-dynamic": "^10.1.2",
-    "@angular/router": "^10.1.2",
+    "@angular/common": "^10.2.0",
+    "@angular/compiler": "^10.2.0",
+    "@angular/core": "^10.2.0",
+    "@angular/forms": "^10.2.0",
+    "@angular/material": "^10.2.5",
+    "@angular/platform-browser": "^10.2.0",
+    "@angular/platform-browser-dynamic": "^10.2.0",
+    "@angular/router": "^10.2.0",
     "angular2-draggable": "^2.3.2",
     "core-js": "^2.6.11",
-    "fast-xml-parser": "^3.17.4",
     "hammerjs": "^2.0.8",
     "minimist": "^1.2.5",
     "ng5-slider": "^1.2.4",
@@ -35,10 +34,10 @@
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1001.2",
-    "@angular/cli": "^10.1.2",
-    "@angular/compiler-cli": "^10.1.2",
-    "@angular/language-service": "^10.1.2",
+    "@angular-devkit/build-angular": "^0.1002.0",
+    "@angular/cli": "^10.2.0",
+    "@angular/compiler-cli": "^10.2.0",
+    "@angular/language-service": "^10.2.0",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.12.62",
@@ -50,6 +49,7 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
+    "prettier": "^2.1.2",
     "protractor": "^7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
