@@ -6,20 +6,27 @@ import { AppRoutingModule } from './app-routing.module';
 import { CoreModule } from '@core/core.module';
 import { Ng5SliderModule } from 'ng5-slider';
 import { AngularDraggableModule } from 'angular2-draggable';
+import { HTTP_INTERCEPTORS } from '@angular/common/http';
+import { NoCacheHeadersInterceptor } from '@shared/interceptors/nocache.interceptor';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-  ],
+  declarations: [AppComponent],
   imports: [
     BrowserModule,
     AppRoutingModule,
     CoreModule.forRoot(),
     Ng5SliderModule,
     AngularDraggableModule,
+    BrowserAnimationsModule,
   ],
-  providers: [],
+  providers: [
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: NoCacheHeadersInterceptor,
+      multi: true,
+    },
+  ],
   bootstrap: [AppComponent],
 })
-export class AppModule {
-}
+export class AppModule {}
