@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
-import { QuestionComponentBase } from '@features/questions/components/question-component-base';
+import { QuestionComponentBaseDirective } from '@features/questions/components/question-component-base.directive';
 import { QuestionMultiple } from '@features/questions/interfaces/question-multiple.interface';
 
 @Component({
@@ -8,7 +8,7 @@ import { QuestionMultiple } from '@features/questions/interfaces/question-multip
   styleUrls: ['./question-multiple.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class QuestionMultipleComponent extends QuestionComponentBase<QuestionMultiple> {
+export class QuestionMultipleComponent extends QuestionComponentBaseDirective<QuestionMultiple> {
   @Input()
   public readonly limit: number;
 }
