@@ -1,4 +1,4 @@
 export enum ModuleType {
   BASE = 'base',
-  ADVANCED = 'advanced'
+  ADVANCED = 'advanced',
 }
