@@ -6,7 +6,7 @@ export interface QuestionHotspot extends QuestionBase {
   text: string;
   image: string;
   answer: Answer<number>;
-  options: { id: number, x: number, y: number, width: number, height: number }[];
+  options: { id: number; x: number; y: number; width: number; height: number }[];
   info?: QuestionInfo;
   title?: string;
 }
