@@ -1,11 +1,18 @@
-import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnDestroy, OnInit } from '@angular/core';
-import { FormControl, FormGroup, Validators } from '@angular/forms';
+import {
+  ChangeDetectionStrategy,
+  ChangeDetectorRef,
+  Component,
+  OnDestroy,
+  OnInit,
+} from '@angular/core';
+import { FormArray, FormControl, FormGroup, Validators } from '@angular/forms';
 import { ActivatedRoute, ParamMap } from '@angular/router';
 import { Observable, Subject } from 'rxjs';
 import { map, takeUntil } from 'rxjs/operators';
 import { QuestionType } from '@features/questions/enums/question-type.enum';
 import { QuestionsService } from '@features/questions/services/questions.service';
 import { QuestionsUnionType } from '@features/questions/types/questions-union.type';
+import { QuestionGroupsChoice } from '@features/questions/interfaces/question-groups-choice.interface';
 
 @Component({
   selector: 'app-questions-page',
@@ -23,12 +30,14 @@ export class QuestionsPageComponent implements OnInit, OnDestroy {
   public readonly question$: Observable<QuestionsUnionType>;
 
   public showTransition: boolean;
+  path: string;
 
   private readonly destroyed$ = new Subject<void>();
-
-  constructor(private route: ActivatedRoute,
-              private cdRef: ChangeDetectorRef,
-              private questionsService: QuestionsService) {
+  constructor(
+    private route: ActivatedRoute,
+    private cdRef: ChangeDetectorRef,
+    private questionsService: QuestionsService,
+  ) {
     this.page$ = this.route.paramMap.pipe(map((params: ParamMap) => +params.get('page')));
     this.index$ = this.page$.pipe(map((page: number) => page - 1));
     this.question$ = this.questionsService.question$;
@@ -36,10 +45,17 @@ export class QuestionsPageComponent implements OnInit, OnDestroy {
   }
 
   ngOnInit(): void {
+    this.path = this.route.snapshot.data.path;
     this.questionsService.attach();
-    this.questionsService.changes$.pipe(takeUntil(this.destroyed$)).subscribe(() => this.cdRef.markForCheck());
+    this.questionsService.changes$
+      .pipe(takeUntil(this.destroyed$))
+      .subscribe(() => this.cdRef.markForCheck());
 
-    this.index$.pipe(takeUntil(this.destroyed$)).subscribe((index: number) => this.questionsService.update(index, this.questions.value));
+    this.index$
+      .pipe(takeUntil(this.destroyed$))
+      .subscribe((index: number) =>
+        this.questionsService.update(index, this.questions.value, this.path),
+      );
 
     this.question$.pipe(takeUntil(this.destroyed$)).subscribe((question: QuestionsUnionType) => {
       this.handleTransition(question);
@@ -54,7 +70,7 @@ export class QuestionsPageComponent implements OnInit, OnDestroy {
   }
 
   finish(): void {
-    this.questionsService.update(null, this.questions.value);
+    this.questionsService.update(null, this.questions.value, this.path);
   }
 
   private handleTransition(question: QuestionsUnionType): void {
@@ -70,6 +86,15 @@ export class QuestionsPageComponent implements OnInit, OnDestroy {
       return;
     }
 
+    if (question.type === QuestionType.GROUPS_CHOICE) {
+      const q = question as QuestionGroupsChoice;
+      const skills = new FormArray([]);
+      q.options.forEach(() => {
+        skills.push(new FormControl('', Validators.required));
+      });
+      this.questions.addControl(key, skills);
+      return;
+    }
 
     this.questions.addControl(key, new FormControl(null, [Validators.required]));
   }
