@@ -1,7 +1,7 @@
-import { NgModule } from '@angular/core';
+import { NgModule, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Title } from '@angular/platform-browser';
-import { ReactiveFormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { DragDropModule } from '@angular/cdk/drag-drop';
 import { ContentModule } from '@content/content.module';
 import { ContentService } from '@content/services/content.service';
@@ -20,15 +20,24 @@ import { QuestionDialogueComponent } from './components/question-dialogue/questi
 import { TransitionComponent } from './components/transition/transition.component';
 import { QuestionSingleComponent } from './components/question-single/question-single.component';
 import { QuestionMultipleComponent } from './components/question-multiple/question-multiple.component';
+import { QuestionMultipleChoiceComponent } from './components/question-multiplechoice/question-multiplechoice.component';
 import { QuestionHotspotComponent } from './components/question-hotspot/question-hotspot.component';
 import { RadioModule } from '../../uikit/radio/radio.module';
 import { ButtonModule } from '../../uikit/button/button.module';
 import { ButtonToggleModule } from '../../uikit/button-toggle/button-toggle.module';
-import nnContent from '@i18n/nynorsk.content.json';
+
 import nbContent from '@i18n/bokmal.content.json';
+import nnContent from '@i18n/nynorsk.content.json';
+
+import nbSystemContent from '@i18n/bokmal.system.json';
+import nnSystemContent from '@i18n/nynorsk.system.json';
+
 import { QuestionSliderComponent } from './components/question-slider/question-slider.component';
 import { Ng5SliderModule } from 'ng5-slider';
 import { QuestionMultipleDiffPointsComponent } from './components/question-multiple-diff-points/question-multiple-diff-points.component';
+import { QuestionGroupsChoiceComponent } from '@features/questions/components/question-groups-choice/question-groups-choice.component';
+import { MatRadioModule } from '@angular/material/radio';
+import { MatButtonModule } from '@angular/material/button';
 
 @NgModule({
   declarations: [
@@ -40,15 +49,18 @@ import { QuestionMultipleDiffPointsComponent } from './components/question-multi
     QuestionCodeComponent,
     QuestionMultipleComponent,
     QuestionDialogueComponent,
+    QuestionMultipleChoiceComponent,
     ResultPageComponent,
     AnswersComponent,
     TransitionComponent,
     QuestionSliderComponent,
     QuestionMultipleDiffPointsComponent,
+    QuestionGroupsChoiceComponent,
   ],
   imports: [
     CommonModule,
     ReactiveFormsModule,
+    FormsModule,
     QuestionsRoutingModule,
     DragDropModule,
     SharedModule,
@@ -57,23 +69,24 @@ import { QuestionMultipleDiffPointsComponent } from './components/question-multi
     ButtonModule,
     ButtonToggleModule,
     Ng5SliderModule,
+    MatRadioModule,
+    MatButtonModule,
   ],
-  providers: [
-    QuestionsService,
-    StatisticsService,
-    StatisticsApiService,
-  ],
+  providers: [QuestionsService, StatisticsService, StatisticsApiService],
 })
 export class QuestionsModule {
-  constructor(contentService: ContentService,
-              statisticsService: StatisticsService,
-              titleService: Title) {
-    contentService.set('nb', nbContent);
-    contentService.set('nn', nnContent);
+  constructor(
+    contentService: ContentService,
+    statisticsService: StatisticsService,
+    titleService: Title,
+  ) {
+    contentService.set('nb', { ...nbContent, ...nbSystemContent });
+    contentService.set('nn', { ...nnContent, ...nnSystemContent });
+
     statisticsService.initUser();
 
-    contentService
-      .changes
-      .subscribe(() => titleService.setTitle(`Kompetanse Norge - ${contentService.get('intro.title')}`));
+    contentService.changes.subscribe(() =>
+      titleService.setTitle(`Kompetanse Norge - ${contentService.get('intro.title')}`),
+    );
   }
 }
