@@ -3,17 +3,8 @@ import { CommonModule } from '@angular/common';
 import { ButtonToggleComponent, ButtonToggleGroupDirective } from './button-toggle';
 
 @NgModule({
-  declarations: [
-    ButtonToggleComponent,
-    ButtonToggleGroupDirective,
-  ],
-  exports: [
-    ButtonToggleComponent,
-    ButtonToggleGroupDirective,
-  ],
-  imports: [
-    CommonModule,
-  ],
+  declarations: [ButtonToggleComponent, ButtonToggleGroupDirective],
+  exports: [ButtonToggleComponent, ButtonToggleGroupDirective],
+  imports: [CommonModule],
 })
-export class ButtonToggleModule {
-}
+export class ButtonToggleModule {}
