@@ -6,7 +6,8 @@ export const environment = {
   production: true,
   qbId_nb: '5qbzulst5d',
   qbId_nn: 'gnkzckfbbp',
-  STATISTICS_API: 'http://vox-wp2.vox.no/dataoppsamling/digitaltest',
+  name: 'digitaltesten',
+  STATISTICS_API: 'https://dataoppsamling.azurewebsites.net/api/',
 };
 
 /*
