@@ -2,7 +2,8 @@ from datetime import datetime
 import secrets
 from flask import request
 from flask.json import jsonify
-from . import app, conn
+from . import app
+from .connect import connection
 
 
 @app.route("/")
@@ -12,16 +13,20 @@ def home():
 
 @app.route("/api/ping")
 def pong():
-    return jsonify({"result": "pong"}), 200
+
+    return (
+        jsonify({"AppSecret": "pong"}),
+        200,
+    )
 
 
-@app.route("/api/submituser", methods=["POST"])
+@app.route("/api/submituser", methods=["GET", "POST"])
 def submitUser():
     token = secrets.token_hex(18)
     now = datetime.utcnow()
     formatted = now.strftime("%Y-%m-%d %H:%M:%S")
 
-    cursor = conn.cursor()
+    cursor, conn = connection()
     cursor.execute("INSERT INTO user(id, created) VALUES (%s, %s)", (token, formatted))
     conn.commit()
     cursor.close()
@@ -29,7 +34,7 @@ def submitUser():
     return jsonify(id=token), 201
 
 
-@app.route("/api/submititem", methods=["POST"])
+@app.route("/api/submititem", methods=["GET", "POST"])
 def submitItem():
 
     id = secrets.token_hex(18)
@@ -37,22 +42,23 @@ def submitItem():
     datecreated = now.strftime("%Y-%m-%d %H:%M:%S")
 
     uid = request.args.get("uid")
-    itemId = request.args.get("itemId")
+    itemid = request.args.get("itemid")
     answer = request.args.get("answer")
-    correct = request.args.get("correct")
     correctanswer = request.args.get("correctanswer")
+    correct = request.args.get("correct")
     time = int(request.args.get("time"))
     totaltime = int(request.args.get("totaltime"))
-    ver = request.args.get("ver")
+    ver = request.args.get("ver", "")
     timeout = request.args.get("timeout")
+    name = request.args.get("name")
 
-    cursor = conn.cursor()
+    cursor, conn = connection()
     cursor.execute(
-        "INSERT INTO itemdata (id, uid, itemid, answer, correct, time, datecreated, correctanswer, totaltime, ver, timeout) VALUES(%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)",
+        "INSERT INTO itemdata (id, uid, itemid, answer, correct, time, datecreated, correctanswer, totaltime, ver, timeout, name) VALUES(%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)",
         (
             id,
             uid,
-            itemId,
+            itemid,
             answer,
             correct,
             time,
@@ -61,6 +67,7 @@ def submitItem():
             totaltime,
             ver,
             timeout,
+            name,
         ),
     )
     conn.commit()
